Enable Grasshopper build by default
